Hide window on deactivation under Wayland
The workaround of the code for problems with some X11 WMs needs `KX11Extras`, which does nothing on Wayland. Hence the patch for Wayland.

On the other hand, at least LabWC doesn't have any problem in this regard, and since Wayland compositors are completely responsible for activation/deactivation, we can consider probable problems as bugs in them.